BookingService/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
psycopg2==2.9.1
requests==2.26.0
pika==1.2.0
python-dotenv==0.19.0 